Trajectory Modeling
A program that prompts the user for a distance to the target, and then exhaustively searches all speeds between 1.0 m/s and 30.0 m/s in increments of 1.0 (so 1.0, 2.0, 3.0, etc. up to 30.0) and all angles between 25.0 and 90.0 in increments of 5.0 (25.0, 30.0, 35.0, 4.0, etc. up to 90.0) to find the best angle and initial speed to hit the target. It will report the best angle, best speed, actual distance traveled, and distance to the target to two decimal places.